+++ src/layout.worker.ts
@@ -1,6 +1,6 @@
+import { connectWorker } from '@reactodia/worker-proxy/protocol';
 import * as cola from 'webcola';
 
-import { connectWorker } from './worker-protocol';
 import type { LayoutGraph, LayoutState } from './diagram/layout';
 import {
     DefaultLayoutOptions, blockingDefaultLayout,